% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{as_epiweek}
\alias{as_epiweek}
\alias{as_epiweek.default}
\alias{as_epiweek.Date}
\alias{as_epiweek.POSIXt}
\alias{as_epiweek.character}
\alias{as_epiweek.factor}
\title{Coerce to a epiweek object}
\usage{
as_epiweek(x, ...)

\method{as_epiweek}{default}(x, ...)

\method{as_epiweek}{Date}(x, ...)

\method{as_epiweek}{POSIXt}(x, ...)

\method{as_epiweek}{character}(x, format, tryFormats = c("\%Y-\%m-\%d", "\%Y/\%m/\%d"), ...)

\method{as_epiweek}{factor}(x, format = NULL, ...)
}
\arguments{
\item{x}{\R object.}

\item{...}{Other values passed to as.Date().}

\item{format}{\verb{[character]}

If NULL, then it attempts to pass character values of the form 'YYYY-Www',
otherwise passed through to \code{as.Date()} (default behaviour).}

\item{tryFormats}{\verb{[character]}

Passed through to \code{as.Date()}}
}
\value{
A \verb{<grates_epiweek>} object.
}
\description{
Generic for conversion to \verb{<grates_epiweek>}
}
\details{
\itemize{
\item Date, POSIXct, and POSIXlt are converted with the timezone respected.
}
}
\examples{
as_epiweek(Sys.Date())
as_epiweek(as.POSIXct("2019-03-04 01:01:01", tz = "America/New_York"))
as_epiweek("2019-05-03")
as_epiweek("2021-W03", format = NULL)

}
\seealso{
\code{epiweek()} and \code{as.Date()}.
}
