% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearweek.R
\name{as_yearweek}
\alias{as_yearweek}
\alias{as_yearweek.default}
\alias{as_yearweek.Date}
\alias{as_yearweek.POSIXt}
\alias{as_yearweek.character}
\alias{as_yearweek.factor}
\title{Coerce to a yearweek object}
\usage{
as_yearweek(x, ...)

\method{as_yearweek}{default}(x, ...)

\method{as_yearweek}{Date}(x, firstday = 1L, ...)

\method{as_yearweek}{POSIXt}(x, firstday = 1L, ...)

\method{as_yearweek}{character}(
  x,
  firstday = 1L,
  format,
  tryFormats = c("\%Y-\%m-\%d", "\%Y/\%m/\%d"),
  ...
)

\method{as_yearweek}{factor}(x, firstday = 1L, format = NULL, ...)
}
\arguments{
\item{x}{\R object.}

\item{...}{Other values passed to as.Date().}

\item{firstday}{\verb{[integer]}

The day the week starts on from 1 (Monday) to 7 (Sunday).}

\item{format}{\verb{[character]}

If NULL, then it attempts to pass character values of the form 'YYYY-Www',
otherwise passed through to \code{as.Date()} (default behaviour).}

\item{tryFormats}{\verb{[character]}

Passed through to \code{as.Date()}}
}
\value{
A \verb{<grates_yearweek>} object.
}
\description{
Generic for conversion to <grates_yearweek>
}
\details{
\itemize{
\item Date, POSIXct, and POSIXlt are converted with the timezone respected.
}
}
\examples{
as_yearweek(Sys.Date())
as_yearweek(as.POSIXct("2019-03-04 01:01:01", tz = "America/New_York"))
as_yearweek("2019-05-03", firstday = 5L)
as_yearweek("2021-W03", format = NULL)

}
\seealso{
\code{as.Date()} and \code{yearweek()}.
}
