% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{clock_locale}
\alias{clock_locale}
\title{clock locale}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{clock}{\code{\link[clock]{clock_locale}}}
}}

