% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epiweek}
\alias{epiweek}
\alias{is_epiweek}
\title{Construct a epiweek object}
\usage{
epiweek(x = integer())

is_epiweek(object)
}
\arguments{
\item{x}{\verb{[integer]}
Vector representing the number of weeks. \code{double} vectors will be converted
via \code{as.integer(floor(x))}.}

\item{object}{An \R object.}
}
\value{
A \verb{<grates_epiweek>} object.
}
\description{
\code{epiweek()} is a constructor for \verb{<grates_epiweek>} objects.
}
\details{
Epiweeks are defined to start on a Sunday and \verb{<grates_epiweek>} objects are
stored as the number of weeks (starting at 0) from the first Sunday after the
Unix Epoch (1970-01-01). That is, the number of seven day periods from
1970-01-04.

Internally they have the same representation as a \verb{<grates_yearweek_sunday>}
object so are akin to an alias but with a marginally more efficient
implementation.
}
\examples{
epiweek(1:10)

}
\seealso{
\code{yearweek()} and \code{isoweek()}.
}
