% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_quarter.R
\name{format.grate_quarter}
\alias{format.grate_quarter}
\title{Print a grate_quarter object}
\usage{
\method{format}{grate_quarter}(x, ...)
}
\arguments{
\item{x}{A grate_quarter object.}

\item{...}{Not currently used.}
}
\description{
Print a grate_quarter object
}
