% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_yearquarter.R
\name{format.grate_yearquarter}
\alias{format.grate_yearquarter}
\title{Print a grate_yearquarter object}
\usage{
\method{format}{grate_yearquarter}(x, ...)
}
\arguments{
\item{x}{A grate_yearquarter object.}

\item{...}{Not currently used.}
}
\description{
Print a grate_yearquarter object
}
