% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_period.R
\name{format.grates_int_period}
\alias{format.grates_int_period}
\title{Format a grates_int_period object}
\usage{
\method{format}{grates_int_period}(x, sep = "to", ...)
}
\arguments{
\item{x}{A grates_int_period object.}

\item{sep}{Where more than one day is grouped with others, \code{sep} is placed
between the upper and lower bounds when printing.}

\item{...}{Not currently used.}

\item{format}{The format to use for the bounds of each grates_int_period entry.}
}
\description{
Format a grates_int_period object
}
