% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{format.grates_month}
\alias{format.grates_month}
\title{Format a grates_month object}
\usage{
\method{format}{grates_month}(x, format = "\%Y-\%b", sep = "to", ...)
}
\arguments{
\item{x}{A grates_month object.}

\item{format}{The format to use for the bounds of each grates_month entry.}

\item{sep}{Where more than one month is grouped with others, \code{sep} is placed
between the upper and lower bounds when printing.}

\item{...}{Not currently used.}
}
\description{
Format a grates_month object
}
