% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{format.grates_period}
\alias{format.grates_period}
\title{Format a grates_period object}
\usage{
\method{format}{grates_period}(x, format = "\%Y-\%m-\%d", sep = "to", ...)
}
\arguments{
\item{x}{A grates_period object.}

\item{format}{The format to use for the bounds of each grates_period entry.}

\item{sep}{Where more than one day is grouped with others, \code{sep} is placed
between the upper and lower bounds when printing.}

\item{...}{Not currently used.}
}
\description{
Format a grates_period object
}
