% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter.R
\name{format.grates_quarter}
\alias{format.grates_quarter}
\title{Format a grates_quarter object}
\usage{
\method{format}{grates_quarter}(x, ...)
}
\arguments{
\item{x}{A grates_quarter object.}

\item{...}{Not currently used.}
}
\description{
Format a grates_quarter object
}
