% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{format.grates_year}
\alias{format.grates_year}
\title{Format a grates_year object}
\usage{
\method{format}{grates_year}(x, ...)
}
\arguments{
\item{x}{A grates_year object.}

\item{...}{Not currently used.}
}
\description{
Format a grates_year object
}
