% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearweek.R
\name{format.grates_yearweek}
\alias{format.grates_yearweek}
\title{Format a grates_yearweek object}
\usage{
\method{format}{grates_yearweek}(x, ...)
}
\arguments{
\item{x}{A grates_yearweek object.}

\item{...}{Not currently used.}
}
\description{
Format a grates_yearweek object
}
