% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interval_duration.R
\name{get_interval_duration}
\alias{get_interval_duration}
\title{The number of days covered}
\usage{
get_interval_duration(x, ...)
}
\arguments{
\item{x}{A grouped date vector.}

\item{...}{Not currently used.}
}
\value{
The number of days covered from the start to end dates for each element in
the input.
}
\description{
Generic utility function for returning the number of days covered by an element
of a grates object.
}
\examples{
# The following should be TRUE
identical(
    get_interval_duration(yearmonth(2020, 1:3)),
    c(31, 29, 31)
)

}
