% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_yrwk.R
\name{get_week_start}
\alias{get_week_start}
\title{Translate user input to the start date of the week}
\usage{
get_week_start(weekday, numeric = TRUE)
}
\arguments{
\item{a}{Weekday specification: E.g. "ISOweek", "MMWRweek", "EPIweek",
"Mon-week", "Tue-week", etc.}
}
\value{
the corresponding weekday
}
\description{
Translate user input to the start date of the week
}
\examples{
get_week_start("ISOweek")
get_week_start("MMWRweek")
get_week_start("EPIweek")

# weeks that start on saturday
get_week_start("Sat-week")
get_week_start("week: Saturday")
get_week_start("2 weeks: Saturday")

}
\keyword{internal}
