% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{grates-defunct}
\alias{grates-defunct}
\alias{int_period}
\alias{as_int_period}
\alias{is_int_period}
\alias{scale_x_grates_int_period}
\title{Defunct functions in package grates}
\usage{
int_period(x = integer(), n = 1L, origin = 0L)

as_int_period(x, ...)

is_int_period(x)

scale_x_grates_int_period(n.breaks = 6, n, origin)
}
\description{
The functions or variables listed here are now defunct, i.e. no longer
available.
}
\details{
\itemize{
\item The \verb{<grates_int_period>} class and it's associated functions were removed
in version 1.0.0.
}
}
\keyword{internal}
