% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grates-package.R
\docType{package}
\name{grates-package}
\alias{grates}
\alias{grates-package}
\title{grates: Grouped Date Classes}
\description{
Provides a coherent interface and implementation for creating grouped date classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.reconverse.org/grates/}
  \item \url{https://github.com/reconverse/grates}
  \item Report bugs at \url{https://github.com/reconverse/grates/issues}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID})

}
\keyword{internal}
