% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{grates_accessors}
\alias{grates_accessors}
\alias{get_n}
\alias{get_n.default}
\alias{get_n.grates_month}
\alias{get_n.grates_period}
\alias{get_n.grates_int_period}
\alias{get_firstday}
\alias{get_firstday.default}
\alias{get_firstday.grates_yearweek}
\alias{get_week}
\alias{get_week.default}
\alias{get_week.grates_yearweek}
\alias{get_quarter}
\alias{get_quarter.default}
\alias{get_quarter.grates_quarter}
\alias{get_year}
\alias{get_year.default}
\alias{get_year.grates_yearweek}
\alias{get_year.grates_quarter}
\alias{get_year.grates_year}
\alias{get_date_range}
\alias{get_date_range.default}
\alias{get_date_range.grates_yearweek}
\alias{get_date_range.grates_month}
\alias{get_date_range.grates_quarter}
\alias{get_date_range.grates_period}
\alias{get_date_range.grates_year}
\alias{get_date_range.grates_int_period}
\title{Grouped date accessors}
\usage{
get_n(x, ...)

\method{get_n}{default}(x, ...)

\method{get_n}{grates_month}(x, ...)

\method{get_n}{grates_period}(x, ...)

\method{get_n}{grates_int_period}(x, ...)

get_firstday(x, ...)

\method{get_firstday}{default}(x, ...)

\method{get_firstday}{grates_yearweek}(x, ...)

get_week(x, ...)

\method{get_week}{default}(x, ...)

\method{get_week}{grates_yearweek}(x, ...)

get_quarter(x, ...)

\method{get_quarter}{default}(x, ...)

\method{get_quarter}{grates_quarter}(x, ...)

get_year(x, ...)

\method{get_year}{default}(x, ...)

\method{get_year}{grates_yearweek}(x, ...)

\method{get_year}{grates_quarter}(x, ...)

\method{get_year}{grates_year}(x, ...)

get_date_range(x, ...)

\method{get_date_range}{default}(x, ...)

\method{get_date_range}{grates_yearweek}(x, ...)

\method{get_date_range}{grates_month}(x, ...)

\method{get_date_range}{grates_quarter}(x, ...)

\method{get_date_range}{grates_period}(x, ...)

\method{get_date_range}{grates_year}(x, ...)

\method{get_date_range}{grates_int_period}(x, ...)
}
\arguments{
\item{x}{A grates object.}

\item{...}{Not currently used.}
}
\value{
\itemize{
\item \code{get_year()}: The corresponding year values as integer for <grates_year>
and <greats_quarter> objects.
\item \code{get_quarter()}: The corresponding quarter values as integer for
<greats_quarter> objects.
}
}
\description{
Generics and methods to work with grouped date objects.
}
