% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_yearweek.R
\name{hidden-scales}
\alias{hidden-scales}
\alias{scale_x_yearweek_1}
\alias{scale_x_yearweek_2}
\alias{scale_x_yearweek_3}
\alias{scale_x_yearweek_4}
\alias{scale_x_yearweek_5}
\alias{scale_x_yearweek_6}
\alias{scale_x_yearweek_7}
\title{Hidden scales}
\usage{
scale_x_yearweek_1(..., n.breaks = 6)

scale_x_yearweek_2(..., n.breaks = 6)

scale_x_yearweek_3(..., n.breaks = 6)

scale_x_yearweek_4(..., n.breaks = 6)

scale_x_yearweek_5(..., n.breaks = 6)

scale_x_yearweek_6(..., n.breaks = 6)

scale_x_yearweek_7(..., n.breaks = 6)
}
\arguments{
\item{...}{Not currently used.}

\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}
}
\value{
A scale for use with ggplot2.
}
\description{
Wrappers around \code{\link[=scale_x_grates_yearweek]{scale_x_grates_yearweek()}} with pre-specified firstday argument.
}
\keyword{internal}
