% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_period.R
\name{int_period}
\alias{int_period}
\alias{is_int_period}
\title{Construct a grates_int_period object}
\usage{
int_period(x = integer(), n = 1L, origin = 0L)

is_int_period(x)
}
\arguments{
\item{x}{Integer vector representing the number of days at the beginning of
the specified grouping.}

\item{n}{An integer indicating the (fixed) number of days used for
grouping; defaults to 1.}

\item{origin}{Day on which the grouping begins (default 0).}
}
\value{
a <grates_int_period> vector.
}
\description{
\code{period()} is a constructor for a <grates_int_period> object
}
\note{
To allow easy comparison between <grates_int_period> object, the \code{origin}
will be stored as it's value modulo \code{n} (i.e \code{origin <- origin \%\% n}).
}
\examples{
int_period(1, n = 2, origin = 1)

}
