% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{is_grate}
\alias{is_grate}
\alias{is_yrwk}
\alias{is_yrmon}
\alias{is_yrqtr}
\alias{is_yr}
\alias{is_period}
\alias{is_int_period}
\title{#' @rdname grate_accessors
#' @export
get_firstdate <- function(x, ...) {
UseMethod("get_firstdate")
}}
\usage{
is_grate(x)

is_yrwk(x)

is_yrmon(x)

is_yrqtr(x)

is_yr(x)

is_period(x)

is_int_period(x)
}
\arguments{
\item{x}{Grouped date object.}
}
\value{
Logical.
}
\description{
#' @rdname grate_accessors
#' @export
get_firstdate.default <- function(x, ...) {
stop(sprintf("Not implemented for class \%s",
paste(class(x), collapse = ", ")))
}
}
\details{
#' @rdname grate_accessors
#' @export
get_firstdate.period <- function(x, ...) {
new_date(attr(x, "firstdate"))
}

#' @rdname grate_accessors
#' @export
get_firstdate.int_period <- function(x, ...) {
attr(x, "firstdate")
}
Is object a grouped date
}
