% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{month}
\alias{month}
\alias{is_month}
\title{Construct a grates_month object}
\usage{
month(x = integer(), n = 1L, origin = 0L)

is_month(x)
}
\arguments{
\item{x}{Integer vector representing the number of months since the Unix
epoch (1970-01-01).}

\item{n}{Number of months that are being grouped by (default 1).}

\item{origin}{Month since the Unix epoch where grouping begins (default 0).}
}
\description{
\code{month()} is a constructor for a <grates_month> object
}
\examples{
month(1)
month(c(4, 7), n = 3, origin = 1)

}
\references{
The algorithm to convert between dates and months relative to the
UNIX Epoch comes from the work of Davis Vaughan in the unreleased
\href{https://github.com/DavisVaughan/datea/}{datea} package.
}
