% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_date.R
\name{new_date}
\alias{new_date}
\title{Date generator}
\usage{
new_date(x = double())
}
\arguments{
\item{x}{A double vector representing the number of days since the UNIX
"epoch", 1970-01-01.}
}
\value{
a (\link{Date}) object.
}
\description{
This function allows the quick creation of Date. It is based on the internal
\code{.Date()} function.
}
\examples{
new_date(0)

}
\keyword{internal}
