% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_period.R
\name{new_int_period}
\alias{new_int_period}
\alias{is_int_period}
\title{Minimal Constructor for an int_period object (Experimental)}
\usage{
new_int_period(x = integer(), n = 1L)

is_int_period(xx)
}
\arguments{
\item{x}{\verb{[integer]}

Vector representing the number of n-integers from 0.

\code{double} vectors will be converted via \code{as.integer(floor(x))}.}

\item{n}{\verb{[integer]}

Number of integers that are being grouped. Must be greater than 0.}

\item{xx}{\R object.}
}
\value{
A \verb{<grates_int_period>} object.
}
\description{
\code{new_int_period()} is a constructor for \verb{<grates_int_period>} objects aimed
at developers.
}
\details{
\code{grates_int_period} objects are stored as the integer number of n-integer
groups from 0. Here n-integer is taken to mean a 'grouping of n consecutive
integers'.
}
\examples{
new_int_period(1:10, 2L)

}
\keyword{internal}
