% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{new_posixct}
\alias{new_posixct}
\title{POSIXct generator}
\usage{
new_posixct(x = double(), tzone = "")
}
\arguments{
\item{x}{A double vector representing the number of seconds since the UNIX
"epoch", 1970-01-01.}

\item{tzone}{A character vector representing the desired time zone.  Defaults
to "" for the local time zone.  Possible values can be found with
\code{\link[=OlsonNames]{OlsonNames()}}.}
}
\value{
a (\link{POSIXct}) object
}
\description{
This function allows the quick creation of Date. It is based on the internal
\code{.Date()} function.
}
\keyword{internal}
