% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearquarter.R
\name{new_yearquarter}
\alias{new_yearquarter}
\alias{is_yearquarter}
\title{Construct a yearquarter object}
\usage{
new_yearquarter(x = integer())

is_yearquarter(xx)
}
\arguments{
\item{x}{\verb{[integer]}

Vector representing the number of quarters.

\code{double} vectors will be converted via \code{as.integer(floor(x))}.}

\item{xx}{\R object.}
}
\value{
A \verb{<grates_yearquarter>} object.
}
\description{
\code{new_yearquarter()} is a constructor for \verb{<grates_yearquarter>} objects aimed
at developers.
}
\details{
\verb{<yearquarter>} objects are stored as the number of quarters (starting at 0)
since the Unix Epoch (1970-01-01).
}
\examples{
new_yearquarter(1:10)

}
