% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_period.R
\name{print.grate_int_period}
\alias{print.grate_int_period}
\alias{format.grate_int_period}
\title{Print a grate_int_period object}
\usage{
\method{print}{grate_int_period}(x, ...)

\method{format}{grate_int_period}(x, ...)
}
\arguments{
\item{x}{A grate_int_period object.}

\item{...}{Not currently used.}
}
\description{
Print a grate_int_period object
}
