% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_month.R
\name{print.grate_month}
\alias{print.grate_month}
\alias{format.grate_month}
\title{Print a grate_month object}
\usage{
\method{print}{grate_month}(x, format = "\%Y-\%b", sep = "to", ...)

\method{format}{grate_month}(x, format = "\%Y-\%b", sep = "to", ...)
}
\arguments{
\item{x}{A grate_month object.}

\item{format}{The format to use for the bounds of each grate_month entry.}

\item{sep}{Where more than one month is grouped with others, \code{sep} is placed
between the upper and lower bounds when printing.}

\item{...}{Not currently used.}
}
\description{
Print a grate_month object
}
