% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_period.R
\name{print.grate_period}
\alias{print.grate_period}
\alias{format.grate_period}
\title{Print a grate_period object}
\usage{
\method{print}{grate_period}(x, format = "\%Y-\%m-\%d", sep = "to", ...)

\method{format}{grate_period}(x, format = "\%Y-\%m-\%d", sep = "to", ...)
}
\arguments{
\item{x}{A grate_period object.}

\item{format}{The format to use for the bounds of each grate_period entry.}

\item{sep}{Where more than one month is grouped with others, \code{sep} is placed
between the upper and lower bounds when printing.}

\item{...}{Not currently used.}
}
\description{
Print a grate_period object
}
