% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_yearmonth.R
\name{print.grate_yearmonth}
\alias{print.grate_yearmonth}
\alias{format.grate_yearmonth}
\title{Print a grate_yearmonth object}
\usage{
\method{print}{grate_yearmonth}(x, format = "\%Y-\%b", ...)

\method{format}{grate_yearmonth}(x, format = "\%Y-\%b", ...)
}
\arguments{
\item{x}{A grate_yearmonth object.}

\item{format}{The format to use for the yearmonth value.}

\item{...}{Not currently used.}
}
\description{
Print a grate_yearmonth object
}
