% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_period.R
\name{print.grates_int_period}
\alias{print.grates_int_period}
\alias{format.grates_int_period}
\title{Print an int_period object}
\usage{
\method{print}{grates_int_period}(x, ...)

\method{format}{grates_int_period}(x, ...)
}
\arguments{
\item{x}{A \verb{<grates_int_period>} object.}

\item{...}{Not currently used.}
}
\description{
Print an int_period object
}
