% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{print.grates_month}
\alias{print.grates_month}
\alias{format.grates_month}
\title{Format and print a month object}
\usage{
\method{print}{grates_month}(x, format = "\%Y-\%b", sep = "to", ...)

\method{format}{grates_month}(x, format = "\%Y-\%b", sep = "to", ...)
}
\arguments{
\item{x}{A \verb{<grates_month>} object.}

\item{format}{\verb{[character]}

The format to use for the bounds of each value.}

\item{sep}{\verb{[character]}

Where more than one month is grouped with others, \code{sep} is placed between the
upper and lower bounds when printing.}

\item{...}{Not currently used.}
}
\value{
For \code{format()}, a character vector representing the formatted input.
\code{print()} is called for the side effect of printing to screen and thus
returns the input \verb{<grates_month>} object invisibly.
}
\description{
Format and print a month object
}
