% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{print.grates_year}
\alias{print.grates_year}
\alias{format.grates_year}
\title{Print a year-quarter object}
\usage{
\method{print}{grates_year}(x, ...)

\method{format}{grates_year}(x, ...)
}
\arguments{
\item{x}{A \verb{<grates_year>} object.}

\item{...}{Not currently used.}
}
\description{
Print a year-quarter object
}
