% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearmonth-class.R
\name{print.grates_yearmonth}
\alias{print.grates_yearmonth}
\alias{format.grates_yearmonth}
\title{Format and print a yearmonth object}
\usage{
\method{print}{grates_yearmonth}(x, format = "\%Y-\%b", ...)

\method{format}{grates_yearmonth}(x, format = "\%Y-\%b", ...)
}
\arguments{
\item{x}{A \verb{<grates_yearmonth>} object.}

\item{format}{\verb{[character]}

The format to use for printing.}

\item{...}{Not currently used.}
}
\value{
For \code{format()}, a character vector representing the formatted input.
\code{print()} is called for the side effect of printing to screen and thus
returns the input \verb{<grates_yearmonth>} object invisibly.
}
\description{
Format and print a yearmonth object
}
