% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearquarter.R
\name{print.grates_yearquarter}
\alias{print.grates_yearquarter}
\alias{format.grates_yearquarter}
\title{Print a year-quarter object}
\usage{
\method{print}{grates_yearquarter}(x, ...)

\method{format}{grates_yearquarter}(x, ...)
}
\arguments{
\item{x}{An \verb{<grates_yearquarter>} object.}

\item{...}{Not currently used.}
}
\description{
Print a year-quarter object
}
