% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarter.R
\name{quarter}
\alias{quarter}
\alias{is_quarter}
\title{Construct a grates_quarter object}
\usage{
quarter(x = integer())

is_quarter(x)
}
\arguments{
\item{x}{Integer vector representing the number of quarters (starting at 0),
since the Unix epoch (1970-01-01).}
}
\description{
\code{quarter()} is a constructor for a <grates_quarter> object
}
\examples{
quarter(0:3)

}
\references{
The algorithm to convert between dates and months relative to the
UNIX Epoch comes from the work of Davis Vaughan in the unreleased
\href{https://github.com/DavisVaughan/datea/}{datea} package.
}
