% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_int_period.R
\name{scale_x_grate_int_period}
\alias{scale_x_grate_int_period}
\title{<grate_period> scale}
\usage{
scale_x_grate_int_period(n.breaks = 6, interval, origin)
}
\arguments{
\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}

\item{interval}{An integer indicating the (fixed) number of days used for
grouping; defaults to 1.}

\item{origin}{The date to anchor the intervals from. All values prior to the
given \code{origin} will be discarded. If NULL (default) the earliest value in
the vector will be used. If specified, \code{origin} must have the same class
as x.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for <grate_period> vector.
}
