% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_quarter.R
\name{scale_x_grate_quarter}
\alias{scale_x_grate_quarter}
\title{<grate_quarter> scale}
\usage{
scale_x_grate_quarter(n.breaks = 6, date_format = NULL)
}
\arguments{
\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}

\item{date_format}{Format to use if "Date" scales are required. If NULL then
labels are centralised and of the form "lower category bound to upper
category bound". If not NULL then the value is used by \code{format.Date()} and
can be any input acceptable by that function.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for <grate_quarter> vector.
}
