% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_yearmonth.R
\name{scale_x_grate_yearmonth}
\alias{scale_x_grate_yearmonth}
\title{<grate_yearmonth> scale}
\usage{
scale_x_grate_yearmonth(
  n.breaks = 6,
  date_format = NULL,
  bounds_format = "\%Y-\%b",
  sep = "to"
)
}
\arguments{
\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}

\item{date_format}{Format to use if "Date" scales are required. If NULL
(default) then labels are centralised and of the form "lower category bound
to upper category bound". If not NULL then the value is used by
\code{format.Date()} and can be any input acceptable by that function.}

\item{bounds_format}{Format to use for grouped date labels. Only used if
\code{date_format} is NULL.}

\item{sep}{Seperator to use for grouped date labels.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for <grate_yearmonth> vector.
}
