% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_yearquarter.R
\name{scale_x_grate_yearquarter}
\alias{scale_x_grate_yearquarter}
\title{<grate_yearquarter> scale}
\usage{
scale_x_grate_yearquarter(n.breaks = 6, date_format = NULL)
}
\arguments{
\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}

\item{date_format}{Format to use if "Date" scales are required. If NULL
(default) then labels are centralised and of the form "lower category bound
to upper category bound". If not NULL then the value is used by
\code{format.Date()} and can be any input acceptable by that function.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for <grate_yearquarter> vector.
}
