% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-int_period.R
\name{scale_x_grates_int_period}
\alias{scale_x_grates_int_period}
\title{int_period scale}
\usage{
scale_x_grates_int_period(
  ...,
  breaks = ggplot2::waiver(),
  n.breaks = 6L,
  centre = FALSE,
  n
)
}
\arguments{
\item{...}{Not currently used.}

\item{breaks}{A \verb{<grates_int_period>} vector of the desired breaks.}

\item{n.breaks}{\verb{[integer]}

Approximate number of breaks calculated using \code{scales::breaks_pretty}
(default 6L).

Will only have an effect if \code{breaks = waiver()}.}

\item{centre}{Only applicable to an \code{int_period} object with \code{n > 1}.

If FALSE labels are place at the edge of the bounds.

If TRUE then labels are centralised and of the form \verb{[lower, upper]}}

\item{n}{\verb{[integer]}

Number used for the original grouping.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for an int_period vector.
}
