% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-isoweek.R
\name{scale_x_grates_isoweek}
\alias{scale_x_grates_isoweek}
\title{isoweek scale}
\usage{
scale_x_grates_isoweek(..., n.breaks = 6, format = NULL)
}
\arguments{
\item{...}{Not currently used.}

\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}

\item{format}{Format to use if "Date" scales are required. If NULL (default)
then labels are in the standard yearweek format (YYYY-Www). If not NULL
then the value is used by \code{format.Date()} and can be any input acceptable
by that function.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for an \verb{<grates_isoweek>} vector.
}
