% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-period.R
\name{scale_x_grates_period}
\alias{scale_x_grates_period}
\title{Period scale}
\usage{
scale_x_grates_period(
  ...,
  breaks = ggplot2::waiver(),
  n.breaks = 6L,
  format = "\%Y-\%m-\%d",
  n,
  offset
)
}
\arguments{
\item{...}{Not currently used.}

\item{breaks}{A \verb{<grates_period>} vector of the desired breaks.}

\item{n.breaks}{\verb{[integer]}

Approximate number of breaks calculated using \code{scales::breaks_pretty}
(default 6L).

Will only have an effect if \code{breaks = waiver()}.}

\item{format}{Format to use for dates.

Value is used by \code{format.Date()} and can be any input acceptable by that
function.}

\item{n}{\verb{[integer]}

Number of days in each period.}

\item{offset}{\verb{[integer]}

Number of days used in original grouping for the offset from the Unix Epoch.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for an \verb{<grates_period>} vector.
}
