% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_quarter.R
\name{scale_x_grates_quarter}
\alias{scale_x_grates_quarter}
\title{<grates_quarter> scale}
\usage{
scale_x_grates_quarter(n.breaks = 6, format = NULL)
}
\arguments{
\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}

\item{format}{Format to use if "Date" scales are required. If NULL then
labels are centralised and of the form "YYYY-Qq. If not NULL then the value
is used by \code{format.Date()} and can be any input acceptable by that function.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for <grates_quarter> vector.
}
