% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_year.R
\name{scale_x_grates_year}
\alias{scale_x_grates_year}
\title{<grates_year> scale}
\usage{
scale_x_grates_year(..., n.breaks = 6)
}
\arguments{
\item{...}{Not currently used.}

\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for <grates_year> vector.
}
