% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_int_period.R
\name{as_int_period}
\alias{as_int_period}
\alias{as_int_period.default}
\alias{as_int_period.int_period}
\alias{as_int_period.integer}
\alias{as_int_period.numeric}
\title{Convert an object to an integer period}
\usage{
as_int_period(x, interval, firstdate, ...)

\method{as_int_period}{default}(x, interval, firstdate, ...)

\method{as_int_period}{int_period}(x, interval = 1L, firstdate = NULL, ...)

\method{as_int_period}{integer}(x, interval = 1L, firstdate = NULL, ...)

\method{as_int_period}{numeric}(x, interval = 1L, firstdate = NULL, ...)
}
\arguments{
\item{x}{An object to coerce to int_period.}

\item{interval}{How many days to include in each period.}

\item{firstdate}{The date to start the intervals from.  If NULL (default) the
earliest date in the vector will be used.}

\item{...}{Not used.}
}
\value{
An \code{int_period} object.
}
\description{
Convert an object to an integer period
}
