% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_period.R
\name{as_period}
\alias{as_period}
\alias{as_period.default}
\alias{as_period.period}
\alias{as_period.Date}
\alias{as_period.POSIXt}
\alias{as_period.character}
\alias{as_period.factor}
\title{Convert an object to a period}
\usage{
as_period(x, interval, firstdate, ...)

\method{as_period}{default}(x, interval, firstdate, ...)

\method{as_period}{period}(x, interval = 1L, firstdate = NULL, ...)

\method{as_period}{Date}(x, interval = 1L, firstdate = NULL, ...)

\method{as_period}{POSIXt}(x, interval = 1L, firstdate = NULL, ...)

\method{as_period}{character}(x, interval = 1L, firstdate = NULL, ...)

\method{as_period}{factor}(x, interval = 1L, firstdate = NULL, ...)
}
\arguments{
\item{x}{`An object to coerce to period.}

\item{interval}{How many days to include in each period.}

\item{firstdate}{The date to firstdate the intervals from.  If NULL (default) the
earliest date in the vector will be used.}

\item{...}{Not used.}
}
\description{
Convert an object to a period
}
