% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{grate-scales}
\alias{grate-scales}
\alias{scale_x_yrwk}
\title{Scales for ggplot2}
\usage{
scale_x_yrwk(..., n = 5, firstday)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}.}

\item{n}{Number of breaks (default 5)}

\item{firstday}{Integer value of the first weekday: 1 (Monday) to 7
(Sunday).}
}
\description{
ggplot2 scales grouped date objects; \code{yrwk}, \code{yrmon}, \code{yrqtr} and \code{period}.
}
