% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{grate_accessors}
\alias{grate_accessors}
\alias{get_week}
\alias{get_week.default}
\alias{get_week.yrwk}
\alias{get_year}
\alias{get_year.default}
\alias{get_year.yrwk}
\alias{get_firstday}
\alias{get_firstday.default}
\alias{get_firstday.yrwk}
\alias{get_interval.yrwk}
\title{Grouped date accessors}
\usage{
get_week(x, ...)

\method{get_week}{default}(x, ...)

\method{get_week}{yrwk}(x, ...)

get_year(x, ...)

\method{get_year}{default}(x, ...)

\method{get_year}{yrwk}(x, ...)

get_firstday(x, ...)

\method{get_firstday}{default}(x, ...)

\method{get_firstday}{yrwk}(x, ...)

get_interval.yrwk(x, days = FALSE, ...)
}
\arguments{
\item{x}{A yrwk, yrmon, yrqtr or period object.}

\item{...}{Not used.}

\item{days}{Should periods be converted in to a number of days.}
}
\description{
Generics and methods to work with grouped date objects:
\itemize{
\item \code{get_week()} returns the corresponding week values for a yrwk vector.
\item \code{get_year()} returns the year.
\item \code{get_firstday()} returns the firstday attribute of a yrwk object.
\item \code{get_month()} returns the month.
}

#' @rdname grate_accessors
#' @export
get_month.default <- function(x, ...) {
stop(sprintf("Not implemented for class \%s",
paste(class(x), collapse = ", ")))
}

#' @rdname grate_accessors
#' @export
get_year.yrqtr <- function(x, ...) {
yrqtr_to_year(x)
}

#' @rdname grate_accessors
#' @export
get_interval.int_period <- function(x, days = FALSE, ...) {
res <- attr(x, "interval")
if (days) {
res <- get_interval_days(x, attr(x, "interval"))
}
res
}
}
\details{
#' @param style Either "numeric" (default) for the integer month value or
#'   "named" to return the abbreviated month name in the current locale.
#'
#' @rdname grate_accessors
#' @export
get_month.yrmon <- function(x, style = c("numeric", "named"), ...) {
yrmon_to_month(x, style = style)
}

#' @rdname grate_accessors
#' @export
get_quarter <- function(x, ...) {
UseMethod("get_quarter")
}

#' @rdname grate_accessors
#' @export
get_quarter.default <- function(x, ...) {
stop(sprintf("Not implemented for class \%s",
paste(class(x), collapse = ", ")))
}

#' @rdname grate_accessors
#' @export
get_quarter.yrqtr <- function(x, ...) {
yrqtr_to_quarter(x)
}

#' @rdname grate_accessors
#' @export
get_year.yr <- function(x, ...) {
unclass(x)
}
}
\examples{
x <- as_yrwk(Sys.Date())
get_year(x)
get_week(x)
get_firstday(x)
}
