% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{grouped_date_accessors}
\alias{grouped_date_accessors}
\alias{get_firstday}
\alias{get_firstday.default}
\alias{get_firstday.grates_yearweek_monday}
\alias{get_firstday.grates_yearweek_tuesday}
\alias{get_firstday.grates_yearweek_wednesday}
\alias{get_firstday.grates_yearweek_thursday}
\alias{get_firstday.grates_yearweek_friday}
\alias{get_firstday.grates_yearweek_saturday}
\alias{get_firstday.grates_yearweek_sunday}
\alias{get_week}
\alias{get_week.default}
\alias{get_week.grates_yearweek}
\alias{get_week.grates_epiweek}
\alias{get_week.grates_isoweek}
\alias{get_year}
\alias{get_year.default}
\alias{get_year.grates_yearweek}
\alias{get_year.grates_epiweek}
\alias{get_year.grates_isoweek}
\alias{get_year.grates_yearmonth}
\alias{get_year.grates_yearquarter}
\alias{get_year.grates_year}
\alias{get_n}
\alias{get_n.default}
\alias{get_n.grates_month}
\alias{get_n.grates_period}
\alias{get_offset}
\alias{get_offset.default}
\alias{get_offset.grates_period}
\title{Accessors for grate objects}
\usage{
get_firstday(x, ...)

\method{get_firstday}{default}(x, ...)

\method{get_firstday}{grates_yearweek_monday}(x, ...)

\method{get_firstday}{grates_yearweek_tuesday}(x, ...)

\method{get_firstday}{grates_yearweek_wednesday}(x, ...)

\method{get_firstday}{grates_yearweek_thursday}(x, ...)

\method{get_firstday}{grates_yearweek_friday}(x, ...)

\method{get_firstday}{grates_yearweek_saturday}(x, ...)

\method{get_firstday}{grates_yearweek_sunday}(x, ...)

get_week(x, ...)

\method{get_week}{default}(x, ...)

\method{get_week}{grates_yearweek}(x, ...)

\method{get_week}{grates_epiweek}(x, ...)

\method{get_week}{grates_isoweek}(x, ...)

get_year(x, ...)

\method{get_year}{default}(x, ...)

\method{get_year}{grates_yearweek}(x, ...)

\method{get_year}{grates_epiweek}(x, ...)

\method{get_year}{grates_isoweek}(x, ...)

\method{get_year}{grates_yearmonth}(x, ...)

\method{get_year}{grates_yearquarter}(x, ...)

\method{get_year}{grates_year}(x, ...)

get_n(x, ...)

\method{get_n}{default}(x, ...)

\method{get_n}{grates_month}(x, ...)

\method{get_n}{grates_period}(x, ...)

get_offset(x, ...)

\method{get_offset}{default}(x, ...)

\method{get_offset}{grates_period}(x, ...)
}
\arguments{
\item{x}{\R object}

\item{...}{Not currently used}
}
\value{
Requested value or an error if no method available.
}
\description{
Generics and methods for accessing information about grouped date objects.
}
\examples{
dates <- as.Date("2020-01-01") + 1:14
dat <- as_isoweek(dates)
get_week(dat)
get_year(dat)

}
