% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{hidden-scales}
\alias{hidden-scales}
\alias{scale_x_yrwk_fd_1}
\alias{scale_x_yrwk_fd_2}
\alias{scale_x_yrwk_fd_3}
\alias{scale_x_yrwk_fd_4}
\alias{scale_x_yrwk_fd_5}
\alias{scale_x_yrwk_fd_6}
\alias{scale_x_yrwk_fd_7}
\title{Hidden scales}
\usage{
scale_x_yrwk_fd_1(..., n = 5)

scale_x_yrwk_fd_2(..., n = 5)

scale_x_yrwk_fd_3(..., n = 5)

scale_x_yrwk_fd_4(..., n = 5)

scale_x_yrwk_fd_5(..., n = 5)

scale_x_yrwk_fd_6(..., n = 5)

scale_x_yrwk_fd_7(..., n = 5)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}.}

\item{n}{Number of breaks (default 5)}
}
\description{
Wrappers around \code{\link[=scale_x_yrwk]{scale_x_yrwk()}} with pre-specified firstday argument.
}
\keyword{internal}
