% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_grate.R
\name{is_grate}
\alias{is_grate}
\alias{is_yearweek}
\alias{is_month}
\alias{is_yearmonth}
\alias{is_yearquarter}
\alias{is_year}
\alias{is_period}
\alias{is_int_period}
\title{Is object a grouped date}
\usage{
is_grate(x)

is_yearweek(x)

is_month(x)

is_yearmonth(x)

is_yearquarter(x)

is_year(x)

is_period(x)

is_int_period(x)
}
\arguments{
\item{x}{Grouped date object.}
}
\value{
Logical.
}
\description{
Is object a grouped date
}
