% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_grate.R
\name{is_grate}
\alias{is_grate}
\title{Is object a grouped date}
\usage{
is_grate(x)
}
\arguments{
\item{x}{Grouped date object.}
}
\value{
Logical.
}
\description{
Is object a grouped date
}
\examples{
is_grate(as_yearweek(Sys.Date()))

}
