% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{is_grate}
\alias{is_grate}
\alias{is_yrwk}
\alias{is_yrmon}
\alias{is_yrqtr}
\alias{is_yr}
\alias{is_period}
\alias{is_int_period}
\title{#' @rdname grate_accessors
#' @export
get_interval.yrmon <- function(x, days = FALSE, ...) {
res <- "1 month"
if (days) {
year <- get_year(x)
month <- get_month(x)
res <- days_in_month(year, month)
}
res
}}
\usage{
is_grate(x)

is_yrwk(x)

is_yrmon(x)

is_yrqtr(x)

is_yr(x)

is_period(x)

is_int_period(x)
}
\arguments{
\item{x}{Grouped date object.}
}
\value{
Logical.
}
\description{
#' @rdname grate_accessors
#' @export
get_interval.yrqtr <- function(x, days = FALSE, ...) {
res <- "1 quarter"
if (days) {
year <- get_year(x)
quarter <- get_quarter(x)
res <- days_in_quarter(year, quarter)
}
res
}
}
\details{
#' @rdname grate_accessors
#' @export
get_interval.yr <- function(x, days = FALSE, ...) {
res <- "1 year"
if (days) {
year <- unclass(x)
res <- 365 + is_leap_year(year)
}
res
}

#' @rdname grate_accessors
#' @export
get_firstdate <- function(x, ...) {
UseMethod("get_firstdate")
}

#' @rdname grate_accessors
#' @export
get_firstdate.default <- function(x, ...) {
stop(sprintf("Not implemented for class \%s",
paste(class(x), collapse = ", ")))
}

#' @rdname grate_accessors
#' @export
get_firstdate.period <- function(x, ...) {
new_date(attr(x, "firstdate"))
}

#' @rdname grate_accessors
#' @export
get_firstdate.int_period <- function(x, ...) {
attr(x, "firstdate")
}
Is object a grouped date
}
