% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{print.grates_month}
\alias{print.grates_month}
\alias{format.grates_month}
\title{Print a month object}
\usage{
\method{print}{grates_month}(x, format = "\%Y-\%b", sep = "to", ...)

\method{format}{grates_month}(x, format = "\%Y-\%b", sep = "to", ...)
}
\arguments{
\item{x}{A \verb{<grates_month>} object.}

\item{format}{\verb{[character]}

The format to use for the bounds of each value.}

\item{sep}{\verb{[character]}

Where more than one month is grouped with others, \code{sep} is placed between the
upper and lower bounds when printing.}

\item{...}{Not currently used.}
}
\description{
Print a month object
}
