% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period-class.R
\name{print.grates_period}
\alias{print.grates_period}
\alias{format.grates_period}
\title{Print a period object}
\usage{
\method{print}{grates_period}(x, format = "\%Y-\%m-\%d", sep = "to", ...)

\method{format}{grates_period}(x, format = "\%Y-\%m-\%d", sep = "to", ...)
}
\arguments{
\item{x}{A \verb{<grates_period>} object.}

\item{format}{\verb{[character]}

The format to use for the bounds of each value.}

\item{sep}{\verb{[character]}

Where more than one day is grouped with others, \code{sep} is placed between the
upper and lower bounds when printing.}

\item{...}{Not currently used.}
}
\description{
Print a period object
}
