% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearmonth.R
\name{print.grates_yearmonth}
\alias{print.grates_yearmonth}
\alias{format.grates_yearmonth}
\title{Print a year-month object}
\usage{
\method{print}{grates_yearmonth}(x, format = "\%Y-\%b", ...)

\method{format}{grates_yearmonth}(x, format = "\%Y-\%b", ...)
}
\arguments{
\item{x}{An \verb{<grates_yearmonth>} object.}

\item{format}{The format to use for printing.}

\item{...}{Not currently used.}
}
\description{
Print a year-month object
}
