% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_int_period.R
\name{scale_x_grates_int_period}
\alias{scale_x_grates_int_period}
\title{<grates_int_period> scale}
\usage{
scale_x_grates_int_period(n.breaks = 6, n, origin)
}
\arguments{
\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}

\item{n}{Number of days used for the original grouping.}

\item{origin}{Original day on which the grouping began.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for <grates_int_period> vector.
}
