% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-month.R
\name{scale_x_grates_month}
\alias{scale_x_grates_month}
\title{month scale}
\usage{
scale_x_grates_month(
  ...,
  n.breaks = 6,
  format = "\%Y-\%m-\%d",
  bounds_format = "\%Y-\%b",
  sep = "to",
  n
)
}
\arguments{
\item{...}{Not currently used.}

\item{n.breaks}{Approximate number of breaks calculated using
\code{scales::breaks_pretty} (default 6).}

\item{format}{Format to use if "Date" scales are required. If NULL then
labels are centralised and of the form "lower category bound to upper
category bound". If not NULL then the value is used by \code{format.Date()} and
can be any input acceptable by that function (defaults to "\%Y-\%m-\%d).}

\item{bounds_format}{Format to use for grouped date labels. Only used if
\code{format} is NULL.}

\item{sep}{Separator to use for grouped date labels.}

\item{n}{Number of months used for the original grouping.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for a month vector.
}
