% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-year.R
\name{scale_x_grates_year}
\alias{scale_x_grates_year}
\title{year scale}
\usage{
scale_x_grates_year(..., n.breaks = 6L, format = NULL)
}
\arguments{
\item{...}{Not currently used.}

\item{n.breaks}{\verb{[integer]}

Approximate number of breaks calculated using \code{scales::breaks_pretty}
(default 6L).}

\item{format}{Format to use if "Date" scales are required.

If not NULL then the value is used by \code{format.Date()} and can be any input
acceptable by that function.}
}
\value{
A scale for use with ggplot2.
}
\description{
ggplot2 scale for year vector.
}
