% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{year}
\alias{year}
\alias{is_year}
\title{Construct a grates_year object}
\usage{
year(x = integer())

is_year(x)
}
\arguments{
\item{x}{Integer vector representing the year.}
}
\description{
\code{year()} is a constructor for a <grates_year> object
}
\examples{
year(2021)

}
