% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{year}
\alias{year}
\alias{is_year}
\title{Construct a year object}
\usage{
year(x = integer())

is_year(object)
}
\arguments{
\item{x}{\verb{[integer]}

Vector representing the years.

\code{double} vectors will be converted via \code{as.integer(floor(x))}.}

\item{object}{\R object.}
}
\value{
A \verb{<grates_year>} object.
}
\description{
\code{year()} is a constructor for \verb{<grates_year>} objects.
}
\examples{
year(2011:2020)

}
